Added new models, item and collection crud
Of DatabaseService:

Renamed getHomeCollectionsOwned to getCollectionsOwned and adjusted such that it receives a $parentCollectionReference for retrieval. The same is done for getHomeCollectionsAuthorized.

getHomeCollections is now renamed to getCollections and accepts $parentCollectionID.

To retrieve the collections for the homepage, simply pass in a null String. To retrieve the collections for some other collection, pass in its ID.

Added getAllEntitiesForCollection, which returns an ArrayList of Entity objects for any collection id provided.

Added createItem.

Added getItemsForCollection.

Other:

Added Entity and Item models.